# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_the-monkeys_freerangenotify_internal_domain_environment.CreateRequest:
    properties:
      app_id:
        type: string
      name:
        enum:
        - development
        - staging
        - production
        type: string
    required:
    - app_id
    - name
    type: object
  github_com_the-monkeys_freerangenotify_internal_domain_environment.PromoteRequest:
    properties:
      resources:
        items:
          type: string
        minItems: 1
        type: array
      source_env_id:
        type: string
      target_env_id:
        type: string
    required:
    - resources
    - source_env_id
    - target_env_id
    type: object
...
  internal_interfaces_http_handlers.registerCustomProviderRequest:
    properties:
      channel:
        maxLength: 50
        minLength: 1
        type: string
      headers:
        additionalProperties:
          type: string
        type: object
      name:
        maxLength: 50
        minLength: 1
        type: string
      webhook_url:
        type: string
    required:
    - channel
    - name
    - webhook_url
    type: object
host: localhost:8080
info:
  contact:
    email: support@freerangenotify.com
    name: API Support
  description: High-performance notification service with multi-channel delivery support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: FreeRangeNotify API
  version: 0.1.9-alpha
paths:
...
  /v1/apps/{id}/environments:
    get:
      description: List all environments for an application
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List environments
      tags:
      - Environments
    post:
      consumes:
      - application/json
      description: Create a new environment (e.g., staging, production) for an application
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Environment creation request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_the-monkeys_freerangenotify_internal_domain_environment.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create an environment
      tags:
      - Environments
  /v1/apps/{id}/environments/{envId}:
    delete:
      description: Permanently remove an environment from an application
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Environment ID
        in: path
        name: envId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete an environment
      tags:
      - Environments
    get:
      description: Retrieve a specific environment by ID
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Environment ID
        in: path
        name: envId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get an environment
      tags:
      - Environments
  /v1/apps/{id}/environments/promote:
    post:
      consumes:
      - application/json
      description: Copy templates and configuration from one environment to another
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Promotion request specifying source and target environments
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_the-monkeys_freerangenotify_internal_domain_environment.PromoteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Promote resources between environments
      tags:
      - Environments
...
  /v1/apps/{id}/providers:
    get:
      description: List all registered custom delivery providers for an application
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List custom providers
      tags:
      - Custom Providers
    post:
      consumes:
      - application/json
      description: Register a new custom delivery provider for an application (admin
        or owner)
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Custom provider registration request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_interfaces_http_handlers.registerCustomProviderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Register a custom provider
      tags:
      - Custom Providers
  /v1/apps/{id}/providers/{provider_id}:
    delete:
      description: Remove a custom delivery provider from an application (admin or
        owner)
      parameters:
      - description: Application ID
        in: path
        name: id
        required: true
        type: string
      - description: Provider ID
        in: path
        name: provider_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Remove a custom provider
      tags:
      - Custom Providers
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    description: Application API key for tenant-facing endpoints
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and your API key
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"